pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "inference-sdk>=0.12.0",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "ollama>=0.6.1",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow<12",
    "pydantic>=2.12.5",
    "pyproj>=3.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "rich>=14.2.0",
    "scikit-image>=0.25.2",
    "shapely>=2.1.2",
    "ultralytics>=8.3.237",
    "uvicorn>=0.38.0",
]
